Fixes #22
If the region is deleted, it is deleted from the regions array. Other functions may still expect the region to exist, causing a NullPointerException when they try to access the array. We can check to see if these values are null and stop the music as we know the region has been removed.